+++ packages/rules-engine/src/components/defaultComponents.ts
@@ -1,16 +1,12 @@
 import { QueryBuilder } from 'react-querybuilder';
 import type { ComponentsRE } from '../types';
 import { ActionElementRE } from './ActionElementRE';
+import { ConditionBuilderHeader, ConditionBuilder, ConditionBuilderBody } from './ConditionBuilder';
+import { ConditionCascade } from './ConditionCascade';
 import { ConsequentBuilder } from './ConsequentBuilder';
 import { ConsequentBuilderBody } from './ConsequentBuilderBody';
 import { ConsequentBuilderHeader } from './ConsequentBuilderHeader';
 import { RulesEngineBuilderHeader } from './RulesEngineBuilderHeader';
-import {
-  ConditionBuilderHeader,
-  RulesEngineConditionBuilder,
-  RulesEngineConditionBuilderBody,
-} from './RulesEngineConditionBuilder';
-import { RulesEngineConditionCascade } from './RulesEngineConditionCascade';
 import { ValueSelectorRE } from './ValueSelectorRE';
 
 /**
@@ -27,9 +23,9 @@ export const defaultComponentsRE: {
   addConsequent: typeof ActionElementRE;
   addCondition: typeof ActionElementRE;
   addSubcondition: typeof ActionElementRE;
-  conditionBuilder: typeof RulesEngineConditionBuilder;
-  conditionBuilderBody: typeof RulesEngineConditionBuilderBody;
-  conditionBuilderCascade: typeof RulesEngineConditionCascade;
+  conditionBuilder: typeof ConditionBuilder;
+  conditionBuilderBody: typeof ConditionBuilderBody;
+  conditionBuilderCascade: typeof ConditionCascade;
   conditionBuilderHeader: typeof ConditionBuilderHeader;
   queryBuilder: typeof QueryBuilder;
   removeConsequent: typeof ActionElementRE;
@@ -45,9 +41,9 @@ export const defaultComponentsRE: {
   addConsequent: ActionElementRE,
   addCondition: ActionElementRE,
   addSubcondition: ActionElementRE,
-  conditionBuilder: RulesEngineConditionBuilder,
-  conditionBuilderBody: RulesEngineConditionBuilderBody,
-  conditionBuilderCascade: RulesEngineConditionCascade,
+  conditionBuilder: ConditionBuilder,
+  conditionBuilderBody: ConditionBuilderBody,
+  conditionBuilderCascade: ConditionCascade,
   conditionBuilderHeader: ConditionBuilderHeader,
   queryBuilder: QueryBuilder,
   removeConsequent: ActionElementRE,